+++ src/ethereum_test_specs/state.py
@@ -29,7 +29,12 @@ from ethereum_test_fixtures.state import (
     FixtureTransaction,
 )
 from ethereum_test_forks import Fork
-from ethereum_test_types import Alloc, Environment, Transaction
+from ethereum_test_types import (
+    Alloc,
+    BlockAccessListExpectation,
+    Environment,
+    Transaction,
+)
 
 from .base import BaseTest, OpMode
 from .blockchain import Block, BlockchainTest, Header
@@ -50,6 +55,7 @@ class StateTest(BaseTest):
     engine_api_error_code: Optional[EngineAPIError] = None
     blockchain_test_header_verify: Optional[Header] = None
     blockchain_test_rlp_modifier: Optional[Header] = None
+    expected_block_access_list: Optional[BlockAccessListExpectation] = None
     chain_id: int = 1
 
     supported_fixture_formats: ClassVar[Sequence[FixtureFormat | LabeledFixtureFormat]] = [
@@ -147,6 +153,7 @@ class StateTest(BaseTest):
             "ommers": [],
             "header_verify": self.blockchain_test_header_verify,
             "rlp_modifier": self.blockchain_test_rlp_modifier,
+            "expected_block_access_list": self.expected_block_access_list,
         }
         if not fork.header_prev_randao_required():
             kwargs["difficulty"] = self.env.difficulty