+++ src/ethereum_test_types/__init__.py
@@ -2,6 +2,16 @@
 
 from .account_types import EOA, Alloc
 from .blob_types import Blob
+from .block_access_list import (
+    BalAccountChange,
+    BalBalanceChange,
+    BalCodeChange,
+    BalNonceChange,
+    BalStorageChange,
+    BalStorageSlot,
+    BlockAccessList,
+    BlockAccessListExpectation,
+)
 from .block_types import (
     Environment,
     EnvironmentDefaults,
@@ -36,7 +46,15 @@ from .utils import Removable, keccak256
 __all__ = (
     "Alloc",
     "AuthorizationTuple",
+    "BalAccountChange",
+    "BalBalanceChange",
+    "BalCodeChange",
+    "BalNonceChange",
+    "BalStorageChange",
+    "BalStorageSlot",
     "Blob",
+    "BlockAccessList",
+    "BlockAccessListExpectation",
     "ChainConfig",
     "ChainConfigDefaults",
     "ConsolidationRequest",